Postgres module tweaks
Move array, enum and geometric instances into their own files.

Add json get and put helper methods to make it easy for consumers to
create their own instances.

Move explain syntax into its own file.
