+++ modules/postgres-circe/src/main/scala/doobie/postgres/circe/Instances.scala
@@ -4,8 +4,6 @@
 
 package doobie.postgres.circe
 
-import cats.Show
-import cats.data.NonEmptyList
 import cats.syntax.either.*
 import cats.syntax.show.*
 import doobie.Get
@@ -13,29 +11,15 @@ import doobie.Put
 import io.circe.*
 import io.circe.jawn.*
 import io.circe.syntax.*
-import org.postgresql.util.PGobject
 
 object Instances {
 
-  private implicit val showPGobject: Show[PGobject] = Show.show(_.getValue.take(250))
-
   trait JsonbInstances {
     implicit val jsonbPut: Put[Json] =
-      Put.Advanced.other[PGobject](
-        NonEmptyList.of("jsonb"),
-      ).tcontramap { a =>
-        val o = new PGobject
-        o.setType("jsonb")
-        o.setValue(a.noSpaces)
-        o
-      }
+      doobie.postgres.instances.json.jsonbPutFromString(_.noSpaces)
 
     implicit val jsonbGet: Get[Json] =
-      Get.Advanced.other[PGobject](
-        NonEmptyList.of("jsonb"),
-      ).temap(a =>
-        parse(a.getValue).leftMap(_.show),
-      )
+      doobie.postgres.instances.json.jsonbGetFromString(parse(_).leftMap(_.show))
 
     def pgEncoderPutT[A: Encoder]: Put[A] =
       Put[Json].tcontramap(_.asJson)
@@ -53,21 +37,10 @@ object Instances {
 
   trait JsonInstances {
     implicit val jsonPut: Put[Json] =
-      Put.Advanced.other[PGobject](
-        NonEmptyList.of("json"),
-      ).tcontramap { a =>
-        val o = new PGobject
-        o.setType("json")
-        o.setValue(a.noSpaces)
-        o
-      }
+      doobie.postgres.instances.json.jsonPutFromString(_.noSpaces)
 
     implicit val jsonGet: Get[Json] =
-      Get.Advanced.other[PGobject](
-        NonEmptyList.of("json"),
-      ).temap(a =>
-        parse(a.getValue).leftMap(_.show),
-      )
+      doobie.postgres.instances.json.jsonGetFromString(parse(_).leftMap(_.show))
 
     def pgEncoderPutT[A: Encoder]: Put[A] =
       Put[Json].tcontramap(_.asJson)
