+++ modules/postgres/src/main/scala/doobie/postgres/hi/largeobject.scala
@@ -4,7 +4,7 @@
 
 package doobie.postgres.hi
 
-import cats.syntax.all.*
+import cats.syntax.apply.*
 import doobie.postgres.PFLO
 import doobie.postgres.free.largeobject.LargeObjectIO
 import doobie.util.io.IOActions
